refactor: update onboarding steps import paths and create centralized onboarding configuration file
